# lib/api-zod/src/generated/types/agentChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentChatRequestMode } from "./agentChatRequestMode";
import type { InteractionInput } from "./interactionInput";

export interface AgentChatRequest {
  message: string;
  formData?: InteractionInput | null;
  /** @nullable */
  existingInteractionId?: string | null;
  mode?: AgentChatRequestMode;
}