# lib/api-zod/src/generated/types/agentDraftResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */
import type { FollowUpSuggestion } from "./followUpSuggestion";
import type { InteractionInput } from "./interactionInput";

export interface AgentDraftResponse {
  draft: InteractionInput;
  missingFields: string[];
  followUpSuggestions: FollowUpSuggestion[];
  summary: string;
  toolTrace: string[];
}