# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */
import type { Hcp } from "./hcp";
import type { RecentActivityItem } from "./recentActivityItem";
import type { SentimentBreakdown } from "./sentimentBreakdown";

export interface DashboardSummary {
  totalInteractions: number;
  weekInteractions: number;
  totalHcps: number;
  pendingFollowUps: number;
  sentimentBreakdown: SentimentBreakdown;
  topHcps: Hcp[];
  recentActivity: RecentActivityItem[];
}