# lib/api-zod/src/generated/types/hcp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */

export interface Hcp {
  id: string;
  name: string;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  territory?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  lastInteractionAt?: Date | null;
  interactionCount: number;
}