# lib/api-zod/src/generated/types/materialsCatalog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialsCatalog {
  materials: string[];
  samples: string[];
}