# lib/api-zod/src/generated/types/sourceMode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-first CRM HCP module API
 * OpenAPI spec version: 0.1.0
 */

export type SourceMode = (typeof SourceMode)[keyof typeof SourceMode];

export const SourceMode = {
  form: "form",
  chat: "chat",
  hybrid: "hybrid",
  edit: "edit",
} as const;